Extend AbstractEntityRepository
